# src/main/resources/api/swagger.yaml
openapi: 3.0.2
info:
  title: Products
  description: 'API to manage products from ManoMano'
  version: 1.0.0
tags:
  - name: Products
    description: Product system
paths:
  /products/{catId}/{minPrice}:
    get:
      tags:
        - Products
      summary: Returns information about products
      operationId: getProducts
      parameters:
        - in: path
          name: catId
          schema:
            type: string
          required: true
        - in: path
          name: minPrice
          schema:
            type: number
          required: true
        - in: query
          name: discountExpDate
          allowEmptyValue: false
          description: Discount expiration date (yyyy-MM-dd HH:mm:ss)
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseList'
              example:
                data:
                  products:
                    - productId: 44789
                      startDate: "2021-01-01 12:45:32"
                      endDate: "2021-01-09 09:07:06"
                      price: 1.56
                      categorizationType: "B2B"
                    - productId: 44789
                      startDate: "2021-12-25 00:00:11"
                      endDate: "2021-12-31 04:09:23"
                      price: 3.44
                      categorizationType: "B2C"
        400:
          description: Malformed syntax of the request params
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                status: 400
                title: Bad Request
                detail: Wrong value in the Expiration Date parameter
        401:
          description: Authentication has failed or not yet been provided
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                status: 401
                title: Unauthorized
                detail: Wrong credentials
        500:
          description: Server encountered an unespected problem
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                status: 500
                title: Internal server error
                detail: Unexpected server error
components:
  schemas:
    ProductResponseList:
      description: List of Products
      type: object
      properties:
        categorizationType:
          type: string
        description:
          type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
    ProductResponse:
      description: Product object returned by the API
      type: object
      properties:
        productId:
          description: Product identifier
          type: string
          format: int64
        startDate:
          description: Discount application start date
          type: string
          format: date-time
        endDate:
          description: Discount application end date
          type: string
          format: date-time
        price:
          description: Final product price
          type: number
          format: double  
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: 'about:blank'
        title:
          type: string
          description: A short, summary of the problem type
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          minimum: 100
          maximum: 600
          exclusiveMaximum: true
        detail:
          type: string
          description: Detail of Problem